Remove references to PlainTextByLineStream constructor that takes FileChannel
The PlainTextByLineStream that takes FileChannel was deprecated. We can remove safely remove it after reviewing internal code that was still using it.

There still deprecated methods that are using the constructors, but it will handled in another issue. After that we are save to permanently remove the deprecated methods from PlainTextByLineStream.

See issue OPENNLP-882